binjo-api/pyproject.toml
[project]
name = "binjo-api"
version = "0.1.0"
description = "BINJO Farm Management API — voice farm diary, AI parsing, PDF export"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "openai>=1.60.0",
    "anthropic>=0.42.0",
    "supabase>=2.0.0",
    "reportlab>=4.0.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    # Phase 3: Receipt image processing
    "Pillow>=10.0.0",
    "pillow-heif>=0.18.0",
    # Phase 3: Financial report charts
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
